git_resource: print warning from git rev-list, if any, in get_patch_count/2
In a situation after `git checkout -b <TAG>`, When
rebar_git_resource:get_patch_count/2 is called, the cloned repo is in
a state where it has a branch and a tag of the same name. This causes
`git rev-list <TAG>..HEAD` to print "warning: refname '<TAG>' is
ambiguous." to stdout, which in turn confuses the list_to_integer at
the end of that function and causes a badarg. Redirecting stderr to
/dev/null is sufficient to deal with this issue.

print the warning from git rev-list, if any